+++ src/buganize/gui/__init__.py
@@ -1,4 +1,0 @@
-from .app import main
-from .constants import APP_ICON_NAME, LAUNCHER_ENV_VARS
-
-__all__ = ["APP_ICON_NAME", "LAUNCHER_ENV_VARS", "main"]